tools/requirements.txt
autopep8==1.5.4
Pillow==8.0.1
pycodestyle==2.6.0
pyperclip==1.8.1
toml==0.10.2
